@@ -14,13 +14,18 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.8.0 <3.0.0"
 
+dependency_overrides:
+  dartx: ^0.3.0
+  xml: ^4.1.0
+  petitparser: ^3.0.0
 dependencies:
   flutter:
     sdk: flutter
+    
   states_rebuilder: ^1.14.2
-  dartz: ^0.8.9
+  dartx: ^0.4.2
   data_connection_checker: ^0.3.4
   get_it: ^3.0.3
   retrofit: any
@@ -37,8 +42,13 @@ dependencies:
   flutter_rounded_date_picker: ^1.0.0
   image_gallery_saver: ^1.1.0
   image_picker: ^0.6.1+4
-  spreadsheet_decoder: ^1.1.0
   printing: ^3.1.0
+  flashy_tab_bar: ^0.0.3
+  excel: ^1.1.4
+  downloads_path_provider: ^0.1.0
+  share_extend: "^1.1.9"
+ 
+  
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -48,9 +58,14 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   retrofit_generator: any
+  flutter_launcher_icons: ^0.7.3
   build_runner: any
   injectable_generator: ^0.2.3
   hive_generator: ^0.7.0+2
+flutter_icons:
+  android: "launcher_icon"
+  ios: true
+  image_path: "assets/images/my_icon.png"
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
