@@ -1,21 +1,17 @@
-import 'dart:io';
-import 'package:path/path.dart';
-
-import 'package:class_room_app/core/constants/color_constant.dart';
-import 'package:class_room_app/feature/attendance/domain/entities/attendance_entity.dart';
-import 'package:class_room_app/feature/attendance/presentation/state/attendance_store.dart';
-import 'package:class_room_app/feature/home/presentation/pages/home.dart';
-import 'package:class_room_app/feature/student/presentation/state/student_store.dart';
-import 'package:class_room_app/feature/subject/presentation/state/subject_store.dart';
-import 'package:class_room_app/injection.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:hive/hive.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:spreadsheet_decoder/spreadsheet_decoder.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
+
+import 'core/constants/color_constant.dart';
+import 'feature/attendance/domain/entities/attendance_entity.dart';
+import 'feature/attendance/presentation/state/attendance_store.dart';
+import 'feature/home/presentation/pages/home.dart';
 import 'feature/student/domain/entities/student_entity.dart';
+import 'feature/student/presentation/state/student_store.dart';
 import 'feature/subject/domain/entities/subject_entity.dart';
+import 'feature/subject/presentation/state/subject_store.dart';
+import 'injection.dart';
 import 'injection.dart' as inject;
 
 void main() async {
